Реализует еще один сложный селектор
