+++ src/pages/main-page/components/offers.tsx/offers.tsx
@@ -2,10 +2,9 @@ import { ShortOfferListType, TypesPageKeys } from '../../../../types/types';
 import Sort from '../sort/sort';
 import CardsList from '../../../../components/cards-list/cards-list';
 import { useAppSelector } from '../../../../hooks';
-import { sortOffers } from './utils';
 import {
   selectCurrentCity,
-  selectCurrentSortKey,
+  selectFilteredSortedOffers,
 } from '../../../../store/selectors';
 
 type OffersProps = {
@@ -18,9 +17,7 @@ function Offers(props: OffersProps): JSX.Element {
   const { offers, typesPage, onCardHover } = props;
 
   const currentCity = useAppSelector(selectCurrentCity);
-
-  const currentSortKey = useAppSelector(selectCurrentSortKey);
-  const sortedOffers = sortOffers(offers, currentSortKey);
+  const sortedOffers = useAppSelector(selectFilteredSortedOffers);
 
   const lastCharacter = offers.length !== 1 ? 's' : '';
