+++ src/utils/utils.ts
@@ -1,8 +1,10 @@
+import { TypesSort } from '../const';
 import {
   CityKeys,
   MapPointsListType,
   OfferType,
   ShortOfferListType,
+  SortTypeKeys,
 } from '../types/types';
 
 const adaptToMap = (
@@ -31,4 +33,17 @@ const filterOffersByCity = (
   city: CityKeys
 ): ShortOfferListType => offers.filter((offer) => offer.city.name === city);
 
-export { adaptToMap,filterOffersByCity };
+const sortBy = {
+  [TypesSort.Popular]: (offers: ShortOfferListType) => [...offers],
+  [TypesSort.HighToLow]: (offers: ShortOfferListType) =>
+    [...offers].sort((prev, next) => next.price - prev.price),
+  [TypesSort.LowToHigh]: (offers: ShortOfferListType) =>
+    [...offers].sort((prev, next) => prev.price - next.price),
+  [TypesSort.Rating]: (offers: ShortOfferListType) =>
+    [...offers].sort((prev, next) => next.rating - prev.rating),
+};
+
+const sortOffers = (offers: ShortOfferListType, key: SortTypeKeys) =>
+  sortBy[TypesSort[key]](offers);
+
+export { adaptToMap, filterOffersByCity, sortOffers };
